+++ modules/constants.js
@@ -2,8 +2,6 @@ export const ADD_GROUP = "ADD_GROUP";
 export const REMOVE_GROUP = "REMOVE_GROUP";
 export const SET_CONJUNCTION = "SET_CONJUNCTION";
 export const SET_NOT = "SET_NOT";
-
-export const ADD_NEW_GROUP = "ADD_NEW_GROUP";
 export const ADD_RULE = "ADD_RULE";
 export const REMOVE_RULE = "REMOVE_RULE";
 export const SET_FIELD = "SET_FIELD";