+++ modules/stores/tree.js
@@ -17,17 +17,17 @@ import {
   getOperatorsForField, getFirstOperator, getWidgetForFieldOp,
   getNewValueForFieldOp
 } from "../utils/ruleUtils";
-import {deepEqual, defaultValue} from "../utils/stuff";
+import {deepEqual, defaultValue, applyToJS} from "../utils/stuff";
 import {validateValue} from "../utils/validation";
-
+import omit from "lodash/omit";
+import mapValues from "lodash/mapValues";
 
 /**
  * @param {object} config
  * @param {Immutable.List} path
  * @param {Immutable.Map} properties
  */
-const addNewGroup = (state, path, properties, config) => {
-  const groupUuid = uuid();
+const addNewGroup = (state, path, groupUuid, properties, config) => {
   const rulesNumber = getTotalRulesCountInTree(state);
   const {maxNumberOfRules} = config.settings;
   const canAddNewRule = !(maxNumberOfRules && (rulesNumber + 1) > maxNumberOfRules);
@@ -413,6 +413,7 @@ const setOperator = (state, path, newOperator, config) => {
  */
 const setValue = (state, path, delta, value, valueType, config, asyncListValues, __isInternal) => {
   const {fieldSeparator, showErrorMessage} = config.settings;
+  let isInternalValueChange;
   const valueSrc = state.getIn(expandTreePath(path, "properties", "valueSrc", delta + "")) || null;
   if (valueSrc === "field" && Array.isArray(value))
     value = value.join(fieldSeparator);
@@ -466,7 +467,7 @@ const setValue = (state, path, delta, value, valueType, config, asyncListValues,
       }
       state = state.setIn(expandTreePath(path, "properties", "value", delta + ""), value);
       state = state.setIn(expandTreePath(path, "properties", "valueType", delta + ""), calculatedValueType);
-      state.__isInternalValueChange = __isInternal && !isLastEmpty && !isLastError;
+      isInternalValueChange = __isInternal && !isLastEmpty && !isLastError;
     }
   }
   if (showErrorMessage) {
@@ -474,10 +475,10 @@ const setValue = (state, path, delta, value, valueType, config, asyncListValues,
   }
   if (__isInternal && (isValid && isLastError || !isValid && !isLastError)) {
     state = state.setIn(expandTreePath(path, "properties", "valueError", delta), validateError);
-    state.__isInternalValueChange = false;
+    isInternalValueChange = false;
   }
-
-  return state;
+  
+  return {tree: state, isInternalValueChange};
 };
 
 /**
@@ -561,6 +562,11 @@ const calculateValueType = (value, valueSrc, config) => {
   return calculatedValueType;
 };
 
+const getField = (state, path) => {
+  const field = state.getIn(expandTreePath(path, "properties", "field")) || null;
+  return field;
+};
+
 const emptyDrag = {
   dragging: {
     id: null,
@@ -575,6 +581,24 @@ const emptyDrag = {
   },
 };
 
+const getActionMeta = (action, state) => {
+  const actionKeysToOmit = [
+    'config', 'asyncListValues', '__isInternal'
+  ];
+  const actionTypesToIgnore = [
+    constants.SET_TREE,
+    constants.SET_DRAG_START,
+    constants.SET_DRAG_PROGRESS,
+    constants.SET_DRAG_END,
+  ];
+  let meta = mapValues(omit(action, actionKeysToOmit), applyToJS);
+  let affectedField = action.path && getField(state.tree, action.path) || action.field;
+  if (affectedField)
+    meta.affectedField = affectedField;
+  if (actionTypesToIgnore.includes(action.type) || action.type.indexOf('@@redux') == 0)
+    meta = null;
+  return meta;
+};
 
 /**
  * @param {Immutable.Map} state
@@ -583,71 +607,111 @@ const emptyDrag = {
 export default (config) => {
   const emptyTree = defaultRoot(config);
   const emptyState = Object.assign({}, {tree: emptyTree}, emptyDrag);
-  const unset = {__isInternalValueChange: undefined};
     
   return (state = emptyState, action) => {
+    const unset = {__isInternalValueChange: undefined, __lastAction: undefined};
+    let set = {};
+    let actionMeta = getActionMeta(action, state);
+
     switch (action.type) {
-    case constants.SET_TREE:
-      return Object.assign({}, state, {...unset}, {tree: action.tree});
+    case constants.SET_TREE: {
+      set.tree = action.tree;
+      break;
+    }
 
-    case constants.ADD_NEW_GROUP:
-      return Object.assign({}, state, {...unset}, {tree: addNewGroup(state.tree, action.path, action.properties, action.config)});
+    case constants.ADD_GROUP: {
+      set.tree = addNewGroup(state.tree, action.path, action.id, action.properties, action.config);
+      break;
+    }
 
-    case constants.ADD_GROUP:
-      return Object.assign({}, state, {...unset}, {tree: addItem(state.tree, action.path, "group", action.id, action.properties, action.config)});
+    case constants.REMOVE_GROUP: {
+      set.tree = removeGroup(state.tree, action.path, action.config);
+      break;
+    }
 
-    case constants.REMOVE_GROUP:
-      return Object.assign({}, state, {...unset}, {tree: removeGroup(state.tree, action.path, action.config)});
+    case constants.ADD_RULE: {
+      set.tree = addItem(state.tree, action.path, "rule", action.id, action.properties, action.config);
+      break;
+    }
 
-    case constants.ADD_RULE:
-      return Object.assign({}, state, {...unset}, {tree: addItem(state.tree, action.path, "rule", action.id, action.properties, action.config)});
+    case constants.REMOVE_RULE: {
+      set.tree = removeRule(state.tree, action.path, action.config);
+      break;
+    }
 
-    case constants.REMOVE_RULE:
-      return Object.assign({}, state, {...unset}, {tree: removeRule(state.tree, action.path, action.config)});
+    case constants.SET_CONJUNCTION: {
+      set.tree = setConjunction(state.tree, action.path, action.conjunction);
+      break;
+    }
 
-    case constants.SET_CONJUNCTION:
-      return Object.assign({}, state, {...unset}, {tree: setConjunction(state.tree, action.path, action.conjunction)});
+    case constants.SET_NOT: {
+      set.tree = setNot(state.tree, action.path, action.not);
+      break;
+    }
 
-    case constants.SET_NOT:
-      return Object.assign({}, state, {...unset}, {tree: setNot(state.tree, action.path, action.not)});
+    case constants.SET_FIELD: {
+      set.tree = setField(state.tree, action.path, action.field, action.config);
+      break;
+    }
 
-    case constants.SET_FIELD:
-      return Object.assign({}, state, {...unset}, {tree: setField(state.tree, action.path, action.field, action.config)});
-
-    case constants.SET_OPERATOR:
-      return Object.assign({}, state, {...unset}, {tree: setOperator(state.tree, action.path, action.operator, action.config)});
+    case constants.SET_OPERATOR: {
+      set.tree = setOperator(state.tree, action.path, action.operator, action.config);
+      break;
+    }
 
     case constants.SET_VALUE: {
-      let set = {};
-      const tree = setValue(state.tree, action.path, action.delta, action.value, action.valueType, action.config, action.asyncListValues, action.__isInternal);
-      if (tree.__isInternalValueChange)
-        set.__isInternalValueChange = true;
-      return Object.assign({}, state, {...unset, ...set}, {tree});
+      const {tree, isInternalValueChange} = setValue(
+        state.tree, action.path, action.delta, action.value, action.valueType, action.config, action.asyncListValues, action.__isInternal
+      );
+      set.__isInternalValueChange = isInternalValueChange;
+      set.tree = tree;
+      break;
     }
 
-    case constants.SET_VALUE_SRC:
-      return Object.assign({}, state, {...unset}, {tree: setValueSrc(state.tree, action.path, action.delta, action.srcKey, action.config)});
-
-    case constants.SET_OPERATOR_OPTION:
-      return Object.assign({}, state, {...unset}, {tree: setOperatorOption(state.tree, action.path, action.name, action.value)});
-
-    case constants.MOVE_ITEM:
-      return Object.assign({}, state, {...unset}, {tree: moveItem(state.tree, action.fromPath, action.toPath, action.placement, action.config)});
-
-
-    case constants.SET_DRAG_START:
-      return Object.assign({}, state, {...unset}, {dragStart: action.dragStart, dragging: action.dragging, mousePos: action.mousePos});
-
-    case constants.SET_DRAG_PROGRESS:
-      return Object.assign({}, state, {...unset}, {mousePos: action.mousePos, dragging: action.dragging});
-
-    case constants.SET_DRAG_END:
-      return Object.assign({}, state, {...unset}, {...emptyDrag, tree: checkEmptyGroups(state.tree, config)});
-
-
-    default:
-      return state;
+    case constants.SET_VALUE_SRC: {
+      set.tree = setValueSrc(state.tree, action.path, action.delta, action.srcKey, action.config);
+      break;
     }
+
+    case constants.SET_OPERATOR_OPTION: {
+      set.tree = setOperatorOption(state.tree, action.path, action.name, action.value);
+      break;
+    }
+
+    case constants.MOVE_ITEM: {
+      set.tree = moveItem(state.tree, action.fromPath, action.toPath, action.placement, action.config);
+      break;
+    }
+
+    case constants.SET_DRAG_START: {
+      set.dragStart = action.dragStart;
+      set.dragging = action.dragging;
+      set.mousePos = action.mousePos;
+      break;
+    }
+
+    case constants.SET_DRAG_PROGRESS: {
+      set.mousePos = action.mousePos;
+      set.dragging = action.dragging;
+      break;
+    }
+
+    case constants.SET_DRAG_END: {
+      set.tree = checkEmptyGroups(state.tree, config);
+      set = { ...set, ...emptyDrag };
+      break;
+    }
+
+    default: {
+      break;
+    }
+    }
+
+    if (actionMeta) {
+      set.__lastAction = actionMeta;
+    }
+    
+    return {...state, ...unset, ...set};
   };
 
 };